db_matching_trait renamed
